refactor: Clean Admin UI Reservation page
- There should be no functional changes, mostly splitting up code to components
